Fix memory safety issues
- Fix double-free in Error::new by using detach() to transfer ownership
- Fix use-after-free in Stream::from(&[u8]) by copying bytes into pool memory
- Fix callback memory management - stop incorrectly using Box::from_raw on batons
- Fix CString lifetime issues in commit(), copy(), and list() functions
- Fix propget hash value type - values are svn_string_t structs, not pointers
- Apply clippy fixes and code formatting

All tests now pass under valgrind with no memory access errors.

+++ src/string.rs
@@ -44,7 +44,7 @@ impl<'pool> BStr<'pool> {
     /// Get bytes as slice
     pub fn as_bytes(&self) -> &[u8] {
         let ptr = unsafe { (*self.ptr).data };
-        let len = unsafe { (*self.ptr).len as usize };
+        let len = unsafe { (*self.ptr).len };
         unsafe { std::slice::from_raw_parts(ptr as *const u8, len) }
     }
